requirements.txt
# requirements.txt

Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
mutagen==1.47.0
pycryptodomex==3.19.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
websockets==12.0
yt-dlp==2023.12.30
https://github.com/kivymd/KivyMD/archive/master.zip